additional popup menu works on frontend for one project
